refactor: remove ccstate-react/experimental from scope-review-modal
Move useCCState and useCommand usage from scope-review-modal.tsx into
the signals/zero-page/settings/connectors.ts file. Scope diff loading
is now triggered automatically when setScopeReviewType$ is called with
a non-null type, eliminating the onRef mount hack and the eslint-disable
comment.

Closes #5804
